bugs

+++ clm2xen.h
@@ -18,7 +18,8 @@ typedef struct mus_xen {
 #define XEN_TO_MUS_ANY(obj) ((mus_any *)((XEN_TO_MUS_XEN(obj))->gen))
 #define MUS_CLM_DEFAULT_TABLE_SIZE 512
 #define MUS_CLM_DEFAULT_FREQUENCY 0.0
-int clm_default_table_size_c(void);
+
+off_t clm_default_table_size_c(void);
 double clm_default_frequency_c(void);
 
 bool mus_xen_p(XEN obj);
